prj_script.sh: when building qemu for xen, always use the latest xen *.deb
When we cross build qemu for xen for arm64, we install the xen *.deb file
so we have the arm64 libraries we need.

To ensure we will build qemu with the same version of xen libary that we
will install, remove any old version before installing the current one.

This avoids having to rebuild the container or doing the same manually.

Signed-off-by: Bill Mills <bill.mills@linaro.org>