demo2a: use local file-system for qemu-ram memory file
When we were using /dev/shm the real qemu (run-B.sh) was crashing.
The file was the right size and owned by me.
Perhaps something strange about /dev/shm fs in a container.

Moving to use the local disk for the file worked fine.

Also delete the memory files before we start and after we finish.

Signed-off-by: Bill Mills <bill.mills@linaro.org>