mixins: for busybox, include both inittabs and choose with symlink
Busybox init won't allow a terminal to be declared that does not exist.
Also console ends up being write only so can't stop the count down etc.

So in the generic mixin add both inittabs we will use and use a symlink
to choose which one will be active.

The default at the global level is hvc0.
Add overrides where needed and get rid of the rest.

Signed-off-by: Bill Mills <bill.mills@linaro.org>