linux: build two version of linux-virtio-msg
Switch the source of the virtio-msg kernel to be my repo and branch.
My branch does not alter defconfig but does provide extra fragments.
Use these config fragments to get needed config and to make the
kernel smaller and faster to build.

For demo1, use older version of linux-virtio-msg

Now that linux-virtio-msg has updated SET_VQUEUE message format, it won't
work with the rust side anymore.  So build an older version with new name
and use that for demo1.

Sigh, more kernel builds :(
This should be short term and we will get everyone aligned on one format.

* refactor common var settings to a new function
* define the new build
* add it to the debian disk
* add the modules to the demo1 guest initramfs
* use the new kernel name kernel in the demo script
* use config mixins for all branches

Signed-off-by: Bill Mills <bill.mills@linaro.org>