build: don't use Linux virt.config
Including this config fragment breaks demo1.
Specifically, turning CONFIG_ARCH_NXP or CONFIG_ARCH_MXC breaks the config
for demo1 and the virtio-i2c in the upstream kernel (base level) does not
enumerate.  It is still broken if these two are turned on again later.

So don't use virt.config but include a lot of the platform disable in the
trim.config fragment.

Signed-off-by: Bill Mills <bill.mills@linaro.org>