images: update zephyr build for saved-images
All zephyr images are saved as one tar.gz of the zephyr dir.
We might not build all zephyr targets which means we cant fetch others
or save the new one.

We always put new built images in build/zephyr and images/zephyr is
a symlink to it.  Fetching leaves a real dir at images/zephyr with
all the files.

Therefore we fetch the whole set if this is the first zephyr build and
move the whole dir to build/zephyr before we add any new one.
If images/zephyr is a symlink, we do nothing.

Signed-off-by: Bill Mills <bill.mills@linaro.org>