build: in worktree setup fetch any tag before using it
If a tag exists upstream but is not on any fetched branch,
it work be fetched by default.  We could fetch all tags
but that is more that we need.  So only in the case that
we are building from a tag, fetch just that tag before
doing the worktree creation.

Signed-off-by: Bill Mills <bill.mills@linaro.org>