multi-qemu: handle tmux server already running
If a tmux server is already running, the environment inherited by the new
session and panes will be that at the time the server was created, not
the environment of the shell that creates the new session.

Therefore don't rely on TEST_QEMU_VARS to pass the var file to the
sub commands.  Instead pass the name of the ENV_FILE on the command line
for each sub command.

This env inheritance issue was also effecting NOW and LOGS and we could
get different values of each in the variance scripts instances.

* get rid of TEST_QEMU_VARS
* pass ENV_FILE as first argument to subcmd
* Don't initialize most env vars are script start, move to main()
* Add NOW, LOGS, LOG, TOTALS, and MODE to the ENV_FILE
* PID_FILE does not really work now but preserve it anyway
* get rid if do nothing MODE check,
  when we add test mode it will be in main anyway
* Add a tmux window name (var was used but never set)
* Eliminate redundant -s in tmux new-session
* Quote all tmux names

Signed-off-by: Bill Mills <bill.mills@linaro.org>