build: build-containers, support lite-build for testing
The demo images are pretty large and it takes a lot of time to save and
transfer the images.

Support a lite-build option for container build testing.
These images are much smaller and faster to save/load/push.
These images will still work but the user needs to do run ./setup the
saved-images will be downloaded on the fly.

Signed-off-by: Bill Mills <bill.mills@linaro.org>