loopback: go back to Viresh's version of lb-vhost-frontend
Viresh made changes so the address of the reserved memory block is not
hard coded into the frontend.  The size still is.

So Viresh's version will work in the demo (w/ smaller total RAM size) so
got back to that version.

Also rearrange things to show my branch names as backup.

Signed-off-by: Bill Mills <bill.mills@linaro.org>