demo2b: use virtio-mmio for disk and net
There is an issue with qemu virtio-pci-{blk,scsi} / Xen 4.19 / and Linux.
With Linux v6.11 this was an oops but not fatal.
With Linux v6.16 this is now fatal.

The demo-xen-ffa does not seem to have an issue but it uses a newer xen.

To avoid the oops/crash with virtio-pci/xen 4.19/linux v6.*,
use virtio-mmio instead.

This moves the ivshmem pci device numbers down by 2:
ivshmem #1 was 0000:00:03.0 now 0000:00:01.0
ivshmem #2 was 0000:00:04.0 now 0000:00:02.0

Change kernel config for linux-upstream to build in virtio-mmio.
This is needed for boot w/o initramfs.

Signed-off-by: Bill Mills <bill.mills@linaro.org>