build: fix util targets so they can be built first
If the devmem2 or util_loopback targets were to first to be built, they
would fail as they were not correctly creating the
images/target/aarch64/debian-12 directory.

Fix this by using the $BASE_DIR we already rely on.
Use it for uio_ivshmem_test also to make it consistent.

Signed-off-by: Bill Mills <bill.mills@linaro.org>