demos: force tmux to use UTF8
With recent versions of the debian:12 upstream container, tmux is not
using UTF8 and the line drawing characters are wrong.

It is not clear what exactly changed, probably something in the new
version of libc6.  (The ncurses and tmux packages have not changed, the
output of env and locale are the same.)

Regardless of the exact change, the old situation was fragile and we
should instead, specify what we want.

Using the -u when starting the tmux server will cause it to use UTF8.
This is the simplest fix and it better than trying to set locale env vars.

Signed-off-by: Bill Mills <bill.mills@linaro.org>